requirements.txt
# Jesse+ - AI增强的加密货币量化交易系统依赖

# Jesse核心框架
jesse>=0.50.0

# AI和机器学习
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# 强化学习
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
yfinance>=0.2.0

# 技术分析
ta-lib>=0.4.0
talib-binary>=0.4.0

# 自然语言处理
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0

# 监控和日志
prometheus-client>=0.17.0
grafana-api>=1.0.0
loguru>=0.7.0

# Web界面
streamlit>=1.28.0
plotly>=5.15.0
dash>=2.11.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 数据库
redis>=4.6.0
pymongo>=4.4.0

# 工具库
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.6.0
schedule>=1.2.0

# 高频交易专用
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# 开发和测试
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0